Seperate concepts into seperate files. One for dealing with the channels, another for the menu bar
