+++ src/routes/menubar.svelte
@@ -1,0 +1,36 @@
+
+
+<div
+  id="menu-bar">
+
+    <img
+    id="menu-svg"
+    src="wii bar final.png"
+    alt="wii bar"
+    />
+
+  
+  
+</div>
+
+<style lang="scss">
+    #menu-svg{
+    position: fixed;
+    height: 20vh;
+    bottom: 0%;
+    width: 100%;
+  }
+
+  #menu-bar{
+    position: fixed;
+    height: 100vh - $channel-view-height;
+    width: 100vw;
+    bottom: 100%; 
+    bottom: 0;
+    display: flex;
+    justify-content: center;
+    align-content: center;
+    overflow: hidden;
+    // background-color: white;
+  }
+</style>
