+++ src/variables.scss
@@ -1,1 +1,3 @@
 /* Variables and mixins declared here will be available in all other SCSS files */
+
+$channel-view-height: 80vh;
